+++ sigma-frontend/next.config.mjs
@@ -4,6 +4,7 @@ const withNextIntl = createNextIntlPlugin("./src/i18n/request.ts");
 
 const nextConfig = {
   experimental: {},
+  output: "standalone",
 };
 
 export default withNextIntl(nextConfig);